+++ lib/ui/components/custom_account_list_tile.dart
@@ -1,14 +1,19 @@
 // Flutter imports:
+import 'package:comics_db_app/core/dark_theme_colors.dart';
+import 'package:comics_db_app/domain/blocs/theme/theme_bloc.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
 
 class CustomAccountListTile extends StatelessWidget {
   const CustomAccountListTile({
     Key? key,
     required this.text,
+    required this.icon,
     required this.onTap,
   }) : super(key: key);
 
   final String text;
+  final IconData icon;
   final GestureTapCallback onTap;
 
   @override
@@ -16,7 +21,11 @@ class CustomAccountListTile extends StatelessWidget {
     return ListTile(
       title: Text(
         text,
-        style: Theme.of(context).textTheme.headlineMedium,
+        style: Theme.of(context).textTheme.displayMedium,
+      ),
+      leading: Icon(
+        icon,
+        color: context.read<ThemeBloc>().isDarkTheme ? Colors.white : DarkThemeColors.kPrimaryColor,
       ),
       onTap: onTap,
     );